+++ src/native/posix/scalar.cpp
